[FIX] renamed Container to IsotopeDistribution, trimmed down classes
