+++ src/openms/include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/IsotopeModel.h
@@ -37,7 +37,7 @@
 #define OPENMS_TRANSFORMATIONS_FEATUREFINDER_ISOTOPEMODEL_H
 
 #include <OpenMS/TRANSFORMATIONS/FEATUREFINDER/InterpolationModel.h>
-#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/Container.h>
+#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/IsotopeDistribution.h>
 
 namespace OpenMS
 {
