+++ src/openms/source/ANALYSIS/ID/AccurateMassSearchEngine.cpp
@@ -34,7 +34,7 @@
 
 #include <OpenMS/ANALYSIS/ID/AccurateMassSearchEngine.h>
 #include <OpenMS/CHEMISTRY/EmpiricalFormula.h>
-#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/Container.h>
+#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/IsotopeDistribution.h>
 #include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/CoarseIsotopeDistribution.h>
 #include <OpenMS/CONCEPT/Constants.h>
 #include <OpenMS/DATASTRUCTURES/ListUtils.h>
@@ -1504,7 +1504,7 @@ namespace OpenMS
     // compute theoretical isotope distribution
     IsotopeDistribution iso_dist(form.getIsotopeDistribution(new CoarseIsotopeDistribution((UInt)common_size)));
     std::vector<double> theoretical_iso_dist;
-    for (IsotopeDistribution::ConstIterator iso_it = iso_dist.begin(); iso_it != iso_dist.end(); ++iso_it)
+    for (auto iso_it = iso_dist.begin(); iso_it != iso_dist.end(); ++iso_it)
     {
       theoretical_iso_dist.push_back(iso_it->getIntensity());
     }
