+++ src/openms/source/CHEMISTRY/ISOTOPEDISTRIBUTION/IsotopeDistribution.cpp
@@ -46,7 +46,7 @@
 #include <OpenMS/CONCEPT/LogStream.h>
 
 #include <OpenMS/CONCEPT/Exception.h>
-#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/Container.h>
+#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/IsotopeDistribution.h>
 #include <OpenMS/CHEMISTRY/EmpiricalFormula.h>
 #include <OpenMS/CHEMISTRY/Element.h>
 
@@ -67,8 +67,6 @@ namespace OpenMS
 
   IsotopeDistribution::~IsotopeDistribution()
   {}
-
-  
 
   IsotopeDistribution & IsotopeDistribution::operator=(const IsotopeDistribution & iso)
   {
@@ -237,5 +235,5 @@ namespace OpenMS
                       });
   }
 
-
 }
+
