+++ src/tests/class_tests/openms/source/Element_test.cpp
@@ -39,7 +39,7 @@
 ///////////////////////////
 
 #include <OpenMS/CHEMISTRY/Element.h>
-#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/Container.h>
+#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/IsotopeDistribution.h>
 #include <OpenMS/DATASTRUCTURES/String.h>
 
 using namespace OpenMS;
