+++ src/tests/class_tests/openms/source/IsotopeDistribution_test.cpp
@@ -36,8 +36,8 @@
 ///////////////////////////
 
 // This one is going to be tested.
+#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/IsotopeDistribution.h>
 #include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/CoarseIsotopeDistribution.h>
-#include <OpenMS/CHEMISTRY/ISOTOPEDISTRIBUTION/Container.h>
 ///////////////////////////
 
 // More headers
